docker-simulation/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core dependencies for Multimodal RAG
boto3>=1.34.0
botocore>=1.34.0
langchain>=0.1.0
langchain-aws>=0.1.0

# PDF Processing and OCR
PyMuPDF>=1.24.0
tabula-py>=2.9.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# Vector Search
faiss-cpu>=1.8.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
